improve MarkdownDisplay
