+++ examples/server/webui/src/components/ChatMessage.tsx
@@ -1,14 +1,8 @@
-import React, { useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { useAppContext } from '../utils/app.context';
 import { Message, PendingMessage } from '../utils/types';
-import { classNames, copyStr } from '../utils/misc';
-import Markdown, { ExtraProps } from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import rehypeHightlight from 'rehype-highlight';
-import rehypeKatex from 'rehype-katex';
-import remarkMath from 'remark-math';
-import remarkBreaks from 'remark-breaks';
-import 'katex/dist/katex.min.css';
+import { classNames } from '../utils/misc';
+import MarkdownDisplay from './MarkdownDisplay';
 
 interface SplitMessage {
   content: PendingMessage['content'];
@@ -233,129 +227,4 @@ export default function ChatMessage({
       )}
     </div>
   );
-}
-
-const Pre: React.ElementType<
-  React.ClassAttributes<HTMLPreElement> &
-    React.HTMLAttributes<HTMLPreElement> &
-    ExtraProps & { origContent: string }
-> = ({ node, origContent, ...props }) => {
-  const startOffset = node?.position?.start.offset ?? 0;
-  const endOffset = node?.position?.end.offset ?? 0;
-
-  const [copied, setCopied] = useState(false);
-  const copiedContent = useMemo(
-    () =>
-      origContent
-        .substring(startOffset, endOffset)
-        .replace(/^```[^\n]+\n/g, '')
-        .replace(/```$/g, ''),
-    [origContent, startOffset, endOffset]
-  );
-
-  if (!node?.position) {
-    return <pre {...props} />;
-  }
-
-  return (
-    <div className="relative my-4">
-      <div
-        className="text-right sticky top-4 mb-2 mr-2 h-0"
-        onClick={() => {
-          copyStr(copiedContent);
-          setCopied(true);
-        }}
-        onMouseLeave={() => setCopied(false)}
-      >
-        <button className="badge btn-mini">
-          {copied ? 'Copied!' : '📋 Copy'}
-        </button>
-      </div>
-      <pre {...props} />
-    </div>
-  );
-};
-
-function MarkdownDisplay({ content }: { content: string }) {
-  const preprocessedContent = useMemo(
-    () => preprocessLaTeX(content),
-    [content]
-  );
-  return (
-    <Markdown
-      remarkPlugins={[remarkGfm, remarkMath, remarkBreaks]}
-      rehypePlugins={[rehypeHightlight, rehypeKatex]}
-      components={{
-        pre: (props) => <Pre {...props} origContent={preprocessedContent} />,
-      }}
-    >
-      {preprocessedContent}
-    </Markdown>
-  );
-}
-
-/**
- * Preprocesses LaTeX content by replacing delimiters and escaping certain characters.
- * Ref: https://github.com/remarkjs/react-markdown/issues/785
- */
-export function preprocessLaTeX(content: string): string {
-  // Step 1: Protect code blocks
-  const codeBlocks: string[] = [];
-  content = content.replace(/(```[\s\S]*?```|`[^`\n]+`)/g, (_, code) => {
-    codeBlocks.push(code);
-    return `<<CODE_BLOCK_${codeBlocks.length - 1}>>`;
-  });
-
-  // Step 2: Protect existing LaTeX expressions
-  const latexExpressions: string[] = [];
-  content = content.replace(
-    /(\$\$[\s\S]*?\$\$|\\\[[\s\S]*?\\\]|\\\(.*?\\\))/g,
-    (match) => {
-      latexExpressions.push(match);
-      return `<<LATEX_${latexExpressions.length - 1}>>`;
-    }
-  );
-
-  // Step 3: Escape dollar signs that are likely currency indicators
-  content = content.replace(/\$(?=\d)/g, '\\$');
-
-  // Step 4: Restore LaTeX expressions
-  content = content.replace(
-    /<<LATEX_(\d+)>>/g,
-    (_, index) => latexExpressions[parseInt(index)]
-  );
-
-  // Step 5: Restore code blocks
-  content = content.replace(
-    /<<CODE_BLOCK_(\d+)>>/g,
-    (_, index) => codeBlocks[parseInt(index)]
-  );
-
-  // Step 6: Apply additional escaping functions
-  content = escapeBrackets(content);
-  content = escapeMhchem(content);
-
-  return content;
-}
-
-function escapeBrackets(text: string) {
-  const pattern =
-    /(```[\S\s]*?```|`.*?`)|\\\[([\S\s]*?[^\\])\\]|\\\((.*?)\\\)/g;
-  return text.replaceAll(
-    pattern,
-    (match, codeBlock, squareBracket, roundBracket) => {
-      if (codeBlock) {
-        return codeBlock;
-      } else if (squareBracket) {
-        return `$$${squareBracket}$$`;
-      } else if (roundBracket) {
-        return `$${roundBracket}$`;
-      }
-      return match;
-    }
-  );
-}
-
-function escapeMhchem(text: string) {
-  return text.replaceAll('$\\ce{', '$\\\\ce{').replaceAll('$\\pu{', '$\\\\pu{');
 }
